Improves the docs on derive
So I've done a bit of work to improve the derive docs.

I do think another improvement may be to copy the style
done in the tokio docs and for some simple examples
show how the generated code looks. However, I think fake would benefit
less from this compared to tokio (how a runtime is setup for an async
application is important to more people). So it may just add unnecessary
noise to the docs.

I guess an issue should be made for the tuple enums - or maybe there's
another way to do it I need to experiment/read-macro-code more